Support ppx-inline-test
This PR allows you to `let%test` wherever and the build won't break. We
can finally start including these tests.

For now, the inline-tests will act directly on the main.exe binary. This
is pretty useless since (1) everything runs in the async monad so we
don't know when the tests are done and (2) we shouldn't need to pass CLI
args to something to run tests.

In another PR, we can start splitting out our code into [a] librar{y,ies}