Keygen at compile-time LIVES
Keys are serialized/deserialized with Bigstring into files. Hashes
are baked into the source code to ensure no tampering (need to move to
Sha256 or something -- right now it's md5).

secure_functionalities was moved to `insecure.ml` in nanobit_base. A new
`key_generation` parameter was added. This determines whether or not
keys are generated at compile-time. When key_generation is insecure,
the prover process will not be happy.

For review:
* Read genkeys.ml very carefully
* Read the camlsnark changes to catch memory bugs
* I attempted to rebase properly, but it's possible I missed some
changes, please double check on that too