Inline tests work again
Inline tests broke when nanobit_base was introduced. Don't ask me how
long it took me to figure out that's where the bug was. We need CI
before I go crazy. Seriously.

As a side-effect of attempting to fix this, I split nanobit into a
library with a simple cli layer as the executable.

I also added some trivial tests in a few places so we can catch
regressions sooner (hopefully).