Strongest-common-ancestor test (#95)
Blockchains have blocks inside of them.

We keep track of the snarked blockchains with a graph-structure where a blockchain points to its predecessor. We need this over a list in case we get blocks out of order.

There is a deepest_path function to extract the "true" blockchain chain. This is what we use in the checks (the shortest deepest true blockchain among the nanobits, and the longest deepest true blockchain among the nanobits are compared)

This PR does not yet support the killing and restarting of nanobitses