Kademlia wrapped from Haskell (#126)
Kademlia runs inside a Haskell process that we spawn from OCaml. This required large changes to our build infrastructure and some changes to the ways integration tests work.
The Haskell code is built via nix, the nix files are generated with the cabal2nix tool.In the docker-dev-environment, everything you need is available. If you have a custom environment make sure you install nix: https://nixos.org/nix/ .
There is a new make target make kademlia that will build the kademlia code for you. You do need nix installed and in your PATH for the make target to work.

Integration tests will handle this for you if you thrash changes.

Nix will give us a reproducible build environment if we pin packages, but we are not pinning packages yet. This should work as long as GHC stays stable, the Haskell libraries we are using are pinned w.r.t the constraints in the kademlia-haskell.cabal file.

The base Dockerfile now prefetches nix dependencies. Put future opam installs at the end of this Dockerfile to prevent having to redownload all our deps over and over again.