Client interacts with daemon
The client has two commands at the moment, `send-txn` and `get-balance`.
Right now, `send-txn` is creating a transaction payload from the
information presented on the CLI, then sending over the daemon who signs
the transaction.

`get-balance` just returns a constant right now.

There are a few placeholder implementations of things, please let me
know if we don't want to land any of those or if we want to make scarier
warnings for the incomplete things. For example, I needed the public_key
compression to not crash to test this end-to-end, but it is very
incorrect.