Protocol and minibit pipe wiring functors
Functors, functors, functors. Minibit_pow (the protocol) is a functor
that hides all implementation details and just shows a clear picture of
how the protocol's state machine advances by one step.

"Minibit" is a functor that hides implementation details of networking
and other data structures, but exposes the miner, ledgers, and specific
named API calls through some magic network.

The idea is we'd apply the minibit functor in main