Created Snark Pool #258
In the Snark Pool, it is required to have a set data structure that
could randomly select a random work . This was achieved in the module
Random_set. Random_set contains an array of keys and a hashtable that
maps a key to it's corresponding index in the array. An element from the
array can be selected at  random. When an element gets added, it gets
added to the dynamic array and the hashtable stores the index of the
element.  When an element gets removed, the index of the element gets
hashed from the hashtable. Then, the deleted element and the last
element in the dynamic array gets switched. Finally, the newly last
element gets  deleted from the array and the table Random_set.
Random_set contains an array of keys and a hashtable that maps a key to
it's corresponding index in the array. An element from the array can
be selected at  random. When an element gets added, it gets added to
the dynamic array and the hashtable stores the index of the element.
When an element gets removed, the index of the element gets hashed from
the hashtable. Then, the deleted element and the last element in the
dynamic array gets switched. Finally, the newly last element gets
deleted from the array and the table

Also, created tests for Random_set