Created Snark Pool #258 (#274)
* Created Snark Pool #258

In the Snark Pool, it is required to have a set data structure that
could randomly select a random work . This was achieved in the module
Random_set. Random_set contains an array of keys and a hashtable that
maps a key to it's corresponding index in the array. An element from the
array can be selected at  random. When an element gets added, it gets
added to the dynamic array and the hashtable stores the index of the
element.  When an element gets removed, the index of the element gets
hashed from the hashtable. Then, the deleted element and the last
element in the dynamic array gets switched. Finally, the newly last
element gets  deleted from the array and the table Random_set.
Random_set contains an array of keys and a hashtable that maps a key to
it's corresponding index in the array. An element from the array can
be selected at  random. When an element gets added, it gets added to
the dynamic array and the hashtable stores the index of the element.
When an element gets removed, the index of the element gets hashed from
the hashtable. Then, the deleted element and the last element in the
dynamic array gets switched. Finally, the newly last element gets
deleted from the array and the table

Also, created tests for Random_set

* Fixed formatting issues

* Made revision to snark pool PR

* Added bin_io interface to snark_pool

In order to add bin_io to snark_pool, it requires random_set to
implement bin_io. This require random_set's dyn_arr to implement
bin_io.  As a result, dyn_array is refactored out of ledger into it's
own file and is shared among different files.

* Renamed types and refactored request_work in snark_pool

* added test for snark-pool

* refactored tests for snark pool

* removed gen and sexp in functors

* reformatted code

* Added gen to snark_pool mli