Groth Maller reduction optimization (#382)
This PR optimizes the reduction in libsnark from rank 1 constraint systems to square constraint systems by avoiding creating an additional constraint when an input constraint has been marked as square. To give a bit more detail, right now the reduction works as follows

```
    /**
     * process R1CS constraints, converting a constraint of the form
     *   \sum a_i x_i * \sum b_i x_i = \sum c_i x_i
     * into two constraints
     *   (\sum (a_i + b_i) x_i)^2 = 4 \sum c_i x_i + x'_i
     *   (\sum (a_i - b_i) x_i)^2 = x'_i
     * where x'_i is an extra variable (a separate one for each original
     * constraint)
     *
     * this adds 2 * cs.num_constraints() constraints
     *   (numbered 0 .. 2 * cs.num_constraints() - 1)
     * and cs.num_constraints() extra variables
     *   (numbered cs.num_variables() + 1 .. cs.num_variables() + cs.num_constraints())
     */
```

We change this so that if a constraint has been explicitly marked as square, then we just pass that constraint through. Otherwise, we do the above.

Remaining is doing the work of setting SNARKy to actually mark some constraints as square. For example, boolean constraints.