Input accumulation outside of libsnark (#532)
The first step of the SNARK verifier is computing a commitment to the input in the following way. The verification key has random group elements g1, g2, …, gn (one for each input). If the input is x1, x2, …, xn, you compute the multiexponentation of the g_i and the x_i. Right now this multiexponentiation is happening inside of libsnark. For reasons having to do with accommodating the new curves we want to have some flexibility in how this works and so I've moved it to snarky