Sok (#533)
This branch implements the signature of knowledge which we use for the transaction snark. It is accomplished by merely hashing the digest of the message to sign into the input of the snark. This is justified as follows.

1. If you have a simulation extractable SNARK construction for a relation `R(x, w)` and `H` is a collision resistant hash function then using that same construction for the relation `R_H'` defined by `R_H(h, w) = exists x. H(x) = h and R(x, w)`. also yields a simulation extractable SNARK construction for `R`.
2. In [GM17](https://eprint.iacr.org/2017/540.pdf), it is showed that if you have a simulation-extractable SNARK construction, and a relation `R`, then a signature of knowledge is essentially a simulation-extractable SNARK for the relation `R'((message, x), w) = R(x, w)`.

Composing these two facts means that hashing the message into the top hash input of the SNARK will yield a signature of knowledge.