Fix ledger builder syncing (#605)
This fixes an issue with how ledger builder synching works. On receiving an external transition for which we don't have the source, we sync to the ledger listed in the ledger builder hash (call it `active_ledger`) and then make sure the parallel scan state serves as a proof that there exists a sequence of transitions taking the ledger with hash mentioned in the blockchain state to `active_ledger`.