Use proper ledger-hash when retreiving via query
I verified this works on Evan's start-stop test branch. Hoping this will
transfer to a proper testnet deployment.

Since this is the 3rd time this same bug has appeared in various
materializations, I'm going to make the type of the ledger-hash
different in blockchain_state and the ledger_builder (see #694)