Proof of stake proposer rewrite (#743)
This rewrite enables the consensus mechanism to control the how the proposer schedules proposals. As part of this, the proof of stake consensus mechanism now schedules proposals only for slots for which the vrf evaluation meets the threshold. The proposer now includes some OTP style asynchronous primitives, which can potentially be abstracted out into their own library for use across other components of the protocol.