coinbase w/ zero fee_excess at the top (#687)
1. The statement of the ledger proof emitted by the parallel scan will have zero fee excess.To achieve this, the total fee excess on every level of the (parallel scan) tree should be zero. For example, fee excess in a tree of depth 3 would be:

                        0
                2               -2
            1      1        -1     -1
         1   1  1   -3      1   1  -1 -1         

In a block, when new transactions are added, the total fee excess (after including fee transfers) is always zero. However, the total fee excess on each level could be a non-zero value if some of the super transactions (transactions, fee transfers for the proofs) added to the tree occupy the "next" tree's leaves. For example: 

                        2
                1                1
           2       -1      -1        2
        1   1  -2    1    1   -2   1    1                  1 -3   _   _   _   _   _   _ 

In this case, the total fee excess at the root of the tree is not zero (the total fee excess in system, however, is zero.) 

This occurs when some of the transactions or fee transfers occupy the next tree's leaves (as a result of varying number of leaves being filled)

To fix this, the diff consisting of transactions and prover fees is split into two parts (pre_diffs) where the first part has transactions and fee transfers occupying only current tree and the second part consisting of transactions and fee transfers that occupy the second tree. If all the transactions and fee transfers fit in the first tree then the second part of the diff will be empty. The example will now look like:

                         0
               1                -1
          2     -1          -1         0
       1   1  -2   1     1    -2    1    -1                1  1  -2  _   _   _   _   _ 



2. A coinbase is included in a block if there is enough work available for one slot
Each coinbase consists of a fee transfer to pay the prover for the work (deducted from the coinbase itself.)

There is however, a case when coinbase could be split into two parts:
When the diff is split into two prediffs and if after  adding transactions, the first prediff has two slots remaining which cannot not accommodate transactions (why? The minimum number of slots required to add a single transaction is three (at worst case number of provers: when each pair of proofs is from a different prover). One slot for the transaction and two slots for fee transfers.), then those slots are filled by splitting the coinbase into two parts. 
If it has one slot, then we simply add one coinbase. It is also possible that the first prediff may have no slots left after adding transactions (For example, when there are three slots and maximum number of provers), in which case, we simply add one coinbase as part of the second prediff.

It could happen that after adding transactions, there is not enough work for the coinbase, in which case, the two prediffs are discarded and a single diff is created with coinbase added before adding any transaction (just like when the diff is not split into two prediffs)

3. full-test: 
  Since coinbase is in effect, proposer (high_balance_pk) is not involved in any transaction (except coinbase). Therefore, transactions are sent from a different account