ledger: fix allocator woes from out-of-order sets
The way set_all_accounts_rooted_at_exn was using `allocate`, it was
allocating new indices starting at wherever the end of the array was,
not at the start of the range that was intended to be set. this wasn't
caught by any of our tests mostly by accident: we didn't happen to
use have set_all_accounts_rooted_at_exn out of order.