fix kademlia line parsing
Reader.read will _not_ read everything written to the pipe
at once, for a variety of reasons, and incorrectly splitting
on newlines on a random read buffer isn't going to give you
the actual lines of the output.

also use trace for kademlia-internal messages, and info for
our informational messages.