Assertion: parallel scan (fixes #658 and #627) (#847)
Traverse the parallel scan state in the order of the statements created. Consider the following tree where a number represents the statements as a result of adding transactions in the corresponding block. For example, 1 stands for the all the statements as a result of transactions added in the first block

                    1
            _                2
      4         _        _       3
    _   _     5  5     _  _    _  _

In this case, for the statements to be chained correctly, the traversal order should be starting from level 0-
1-2-3-4-5-5

This could explain why the assertion in coda.ml failed sometimes. Currently, in master, the traversal order in the above example would be 1-2-4-3-5-5