Materialize snarked account db (#925) fixes #339 fixes #869
1. keeping track of the transactions that needs to be undone in order to get to the latest snarked account db
For each level on the tree, a list of transactions corresponding to the statements is stored and when a proof is emitted, the transaction list corresponding to the root is discarded. At any given point, the latest snarked accounts db can be obtained by undoing the transactions stored for each of the levels. (fixes #869)
2. Also using this to validate aux from the network during catch-up (fixes #339)