No more monkey-patches ledger-builder-controller
Fixes #934

I had to introduce one tiny monkey patch in Coda_main for Public_key
because I didn't want this PR to be a trillion lines. I also moved a few
of the interfaces to Coda_pow in an effort to make the future Coda_spec
refactor a bit easier.