Select `coda-kademlia` binary (#965)
As suggested by @jkrauska, kademlia now looks for `coda-kademlia` in
PATH and can be overridden by CODA_KADEMLIA_PATH. There is a nice error
message suggesting you set this environment variable if there is a
failure to find this binary. I kept the other paths as fallback to keep
existing behavior.