Curve abstraction (#551)
This PR makes it possible to use bigger curves. Some other features crept in as well:

- Move common random string generation into its own `crs` library
- Generate sample keypairs at compile time instead of hard-coding them

The most interesting part is the module `Wrap_input` in `lib/crypto_params/crypto_params_init/crypto_params_init.ml`. This specifies how we represent a pedersen hash (a Tick field element) as an input to a tock wrapping SNARK. Essentially we use either one or Tock elements to represent one Tick element depending on whether the Tock field is larger than the Tick field or not.