Make it possible to run Proof-of-signature and Proof-of-stake in devnets/testnets (#1080)
Proof-of-signature and Proof-of-stake both require a private key when proposing. In this PR, we switch to demanding this private key when proposal is enabled. As such we no longer need signer_key for proof-of-signature.

I had to change the Consensus_transition_data for proof-of-signature to be a variant as we don't have the private key readily available for the genesis state. When it's genesis we just store (1,1) as the signature and ignore the signature verification iff this is the genesis transition. I believe that the SNARK circuit is still correct, but please review this part carefully.

The rest of the code just adjusts to the new lack of having this keypair available always and is relatively straightforward.

This fixes the issue we were seeing where the proposer on a devnet would die instantly (as our code was trying to read a secret key out of the genesis ledger which we no longer compile in in production-mode).

Closes #1052