speed up sync ledger answering (#1114)
we spend a surprising amount of time looking up
the local ledger for that hash. for now, cache
the most recent one you saw. this is good enough
to speed up the tests, and almost all cases with
posig (not postake).

tested by looking at a trace view (from my in-progress
tracer) and seeing that the slow bits went away.