Generalize payment to user command (#1122)
This generalizes payment to "user command" in anticipation of adding a stake delegation transaction. A user command is a signed payload issued by a user which contains a fee, sender, and a nonce (as it modifies that sender's account). This PR basically just shuffles around where data is in the payment type and changes some names.