Async tracing (#1121)
This is the first cut of lightweight tracing that uses some
hacked in hooks into async. To use this:

1. `git clone -b tracing https://github.com/CodaProtocol/async_kernel.git`
2. `opam pin add async_kernel`
3. set `tracing` to `true` in your `config.mlh`.
4. build and run some coda
5. `cd src/app/trace-tool`
6. `cargo run --release /tmp/coda-profile-* > profile.json`
7. Load `profile.json` in `chrome://tracing`

This still needs some work, in particular around how it handles "nested" `trace_task`. Right now it treats each of those as a thread, which is the right behavior for the initial class of things I wanted to trace, but doesn't work very well for tasks that get repeated. trace-viewer supports what I need, I just need to emit slightly different things. This can all happen on the trace-tool side. 

trace-tool itself is written in Rust because my early trace dumps were hundreds of megs. it's a bit messy.

This works and is useful already for relatively short runs (like the integration tests). It's important to note that this is all based around _tracing_: emitting a stream of events, as they happen. The rate right now is around 810kB/s.