Fix full-test with-snark bugs (#1115)
## Changes
I rolled back the:

```ocaml
[`Signer of signature | `Genesis]
```

change, and went back to global-signer-key. I increased the `CODA_PROPOSAL_INTERVAL` for full-test with snark as well.

I also fixed various other small issues and made a half-working parallel-scan-state visualizer (that still was helpful in debugging).

Note: full-test passed in the `test_snark` profile on my machine (inside a docker container on a VM in a macbook pro) with the timeouts included here: (15min for part one, and 5min for part two -- see `full_test.ml`).

## Test Plan
I successfully built with the testnet profile with the secret global private signer key (and the snark successfully was created at compile time). We still will want to check this in a devnet as well.

This obsoletes #1104