[RFC] Transition-Frontier-Controller (#1168)
Refactoring the existing ledger-builder-controller to a new simpler set of
components we're calling the transition-frontier-controller. At the top level,
we'll connect these components together in a similar manner to coda_lib --
just wiring together pipes. The new transition-frontier-controller includes the
merkle-mask and a safer history-catchup mechanism in addition to simplifying
asynchronous logic. The goal is to end up with a simple set of components that
will be robust towards future requirements changing and easy to trace dataflow
and test and debug.