Improve Pedersen hash efficiency by using unsafe add (#1315)
This PR shaves 1 constraint off of the curve addition used in the Pedersen hash. Whereas before hashing cost (2 + 4) constraints per triple, it will now cost (2 + 3) constraints per triple. Concretely this should be about a 12% improvement in hashing inside the SNARK.

Doing this though requires introducing a scary function. That function is `add_unsafe` which is a checked function for adding two curve points. The scary part is in `div_unsafe`. Which is a checked function for dividing one field element by another. In the event that both field elements are zero however, the result is totally unconstrained and could be chosen by the prover to be arbitrary.

This then causes `add_unsafe` to have the following behavior. 

- If `p = - q` then `add_unsafe p q` makes an unsatisfiable assertion.
- If `p = q` then `add_unsafe p q` could be many different curve points(!)
- If `p !=  +/- q`, then `add_unsafe p q` is genuinely the curve point `p + q`

Thus, `add_unsafe` must never be called on two potentially identical curve points. All uses of it are in the context of Pedersen hashing in which we add curve points of the form

```
add_unsafe (a_i * g_i + ... + a_j + g_j)  (a_k * g_k + ... + a_l + g_l)
```

with the intervals `[i, j]` and `[k, l]` being disjoint and
where the `g_r` are the random pedersen parameters and `a_s in {+/- 1, …, +/- 4}`.

I claim that the usual argument indicates that one should not be able to find `a_s` causing those two sums to be identical unless one can compute discrete logs in `G`.