Lift snarky types out of the functor (#1379)
This PR lifts the main snarky types out of the snarky functor so that typs and checked functions can be reused across multiple proof systems with the same field. This is necessary for a snarkitecture change I'm making where we use Groth16 in some places and the BG snark in others.