Blog styling and initial blog post on fast accumulation (#1359)
All blog posts will now require an author field. Subtitle and author_website is optional.

There are some shortcuts in here that we'll have to undo when we have more than one blog post on the website (for example the link to the blog in disqus is hardcoded).

The outer-frame of the blog post with the title/subtitle/date/author and share-sheet/disqus at the bottom is in blog_post.ml

All the blog content styling is in blog.css and can't be done with tachyons since the output is generated by pandoc.

There is a static/blog directory for static assets connected to a particular blogpost (new subdir per post).

Also now we prefer katex to mathjax as it's faster and has more features.