Remove gamma from the groth16 snark (#1383)
This PR removes gamma from the Groth16 implementation (or put another way, it sets gamma equal to one). The benefit is that it makes the verification key one G2 element smaller, which is a big win from the point of view of proof composition.

I opened a PR on libsnark [here](https://github.com/scipr-lab/libsnark/pull/137) so hopefully that eventually makes it in upstream, but I want to get this through here first.