change the type of ledger builder diff (#1396)
This is the first of all the PRs related to the refactoring of the create_diff function in Ledger_builder. None of these PRs will compile because they don't have all the changes. They will be merged to a different branch and then to master. It is split this way to make it relatively easy to review.

>Changes

Coinbase in the diff now just has the fee transfer (instead of Completed_work.t) for the work required to include it in the diff. This is then subtracted from the fee transfers for the work required to add user commands.