Create diff refactoring (#1408)
This PR consists of the refactoring change made to create diff. 
The new algorithm first fetches all the completed work and filters valid transactions. These are then checked against three main constraints:
1. Static work constraint: based on maximum number of jobs existing in the system and space available
2. Dynamic work constraint: based on a certain set of user commands and completed work
3. Space constraint: if everything fits in the given space
4. Budget constraint: if the payment fees can pay for the proofs

if any of the constraints do not satisfy, then either a user command or a proof is discarded (See `check_constraints_and_update` function). This is sort of a bottom up approach if we call the previous one top down where transactions were added one by one if proofs were purchased.

Changed the type of `Staged_ledger_diff` in a way that coinbase in the diff now just has the fee transfer (instead of `Transaction_snark_work.t`) for the work required to include it in the diff. This is then subtracted from the fee transfers for the work required to add user commands.

Some benefits of this change:
1. Since these constraint definitions are separated from the process of including the user commands and completed work, we can make changes to them easily in the future
2. No transaction undos

This also has a change related to the FIFO work order (#1345) in which the container capacity is fetched from `Transaction_snark_scan_state` instead of doing the math everywhere

Links to the PRs that were merged into this one: #1396 #1397 #1398 #1399