MacOS supported fork of rpc_parallel (#1418)
macOS doesn't have a procfs. The canonical way to get the full
executable path is to use the C-function _NSGetExecutablePath. I'm
branching on OS using uname -s and using _NSGetExecutablePath on
Darwin systems and keeping the default proc-lookup on other ones.

See src/util.ml for the big change that I made. The rest is just type-error fixups.

This is necessary for builds on macOS to not die after immediately starting up the coda daemon. After these changes, the build doesn't die (amongst other necessary changes not included in this PR).

See janestreet/rpc_parallel#6 for a version of these patches on the master branch of rpc_parallel.