Libsnark get_evaluation_domain works on macOS (#1419)
Before this change on macOS systems, libsnark would crash on
get_root_of_unity and exception handlers would not work. The
hypothesis is that using exceptions for control flow in
get_evaluation_domain triggers some sort of UB (probably because this
runs statically before the main function). On Linux, the UB happens to
work properly, but on OSX it doesn't. As such, I rewrote all the
control-flow related exception handling in get_evaluation_domain to use
an error bool pointer.

I'll open a PR against libsnark directly when the review for this PR
finishes.

On macOS the Coda build succeeds only after these changes are made (as
we use libsnark at compile time). CI will tell us that the build still
works on Linux.