Scan-state-tree view (#1434)
Adding a cli command `snark-job-list` to view the snark jobs (along with the position) on the tree that's in the scan state of the best tip. These are the jobs that are yet to be completed by the snark workers. 
The command outputs a list of `(position,job-view)`. The view of a job here is a record consisting of  source ledger hash, target ledger hash, fee excess, and supply increase. It represents a transition of the ledger state after applying a set of transactions.
Fee excess and supply increase tell us about the net transaction fees and new money minted in the system respectively.

The output of this command is in JSON format to enable use of off-the-shelf tools.

`B` represents a base job and has one record that represents a transition from `source` to `target` as a result of applying a transaction.
`M` represents a merge job that has two records, say record1 and record2, and represents the transition from `record1.source` to `record2.target` as a result of combining two transitions. it could be two `B`s or two `M`s.

For example, output of the command on a tree with transaction-capacity=4 and one base job as a result of enqueuing a coinbase is: 

     $ _build/install/default/bin/coda client snark-job-list
     [[0,{"M":[[],[]]}],[1,{"M":[[],[]]}],[2,{"M":[[],[]]}],[3,{"M":[[],[]]}],[4,{"M":[[],[]]}],[5,{"M":[[],[]]}],[6,{"M":[[],[]]}],[7,{"M":[[],[]]}],[8,{"M":[[],[]]}],[9,{"M":[[],[]]}],[10,{"M":[[],[]]}],[11,{"M":[[],[]]}],[12,{"M":[[],[]]}],[13,{"M":[[],[]]}],[14,{"M":[[],[]]}],[15,{"B":[{"Source":"72424171276081751598787942910321842308858254595640291393753393954111564565245356960661379","Target":"282766584673677276000425149291497796937616006651301184746353531791045773822966800310738233","Fee Excess":{"magnitude":0,"sgn":["Pos"]},"Supply Increase":10}]}],[16,{"B":[[]]}],[17,{"B":[[]]}],[18,{"B":[[]]}],[19,{"B":[[]]}],[20,{"B":[[]]}],[21,{"B":[[]]}],[22,{"B":[[]]}],[23,{"B":[[]]}],[24,{"B":[[]]}],[25,{"B":[[]]}],[26,{"B":[[]]}],[27,{"B":[[]]}],[28,{"B":[[]]}],[29,{"B":[[]]}],[30,{"B":[[]]}]]


The jobs are indexed according to their positions on the tree. In the above example, the scan state tree looks like this:


                                                  0
                            1                                           2
                 3                     4                       5                     6
           7          8          9             10          11          12          13         14
      15     16    17    18    19    20    21    22    23    24    25    26    27    28    29    30 

fixes:  #1294