Build_with_snark does sig and stake (#1484)
We need to make sure to test build with sig and stake SNARKs on CI

In order to fix the stake build, the genesis block considers account zero to be the VRF winner. In addition, there were some missing handlers in the precomputed value case and some rogue functor instantiations.