Fake hashes (#1519)
This allows slotting in SHA256 instead of Pedersen for hashing. It's a pretty gross hack, but it can make for significantly faster integration tests. Before this change, the latency between starting to make a block and broadcasting it was 953ms. With fake_hash true, it is 111ms. There's still some low hanging fruit to bring that down even more.