Transition the frontier root properly updating snarked ledger without leaking memory (#1567)
Refactored system for handling root changes in transition frontier. Merkle 
masks now alias the same record when they are attached and detached and there 
is a remove_and_reparent function for shuffling around chained masks.

See the comments above `add_breadcrumb_exn` for more details.

Additionally, made scan-state emit the transactions pertaining to the ledger-proof whenever a ledger-proof is emitted. This information is already present, but required some juggling of `fst` and `snd` and manipulating signatures.

Also, replaced several assertions with `%test_result` invocations and added visualizers for transition_frontier and merkle masks that were used during debugging (and may be useful in the future).

Closes #1542
Closes #1544