Feature/blockchain snark is groth snark (#1537)
This completes the long series of PRs related to changing the snark architecture. It replaces all non-wrap SNARKs (the step blockchain SNARK, the merge and base transaction SNARKs) with Groth16 rather than Groth--Maller17. This is fine as per conversations with @vanishreerao, although for reasons having to do with the MPC, we may avoid using GrothMaller17 entirely.