Scan-state constants -> compile-time (#1630)
These need to be static for an entire testnet cluster so it makes sense
to make these compile time values. There is a new scan_state directory
in config that has the two constants: transaction_capacity_log_2 and
work_delay_factor.

It compiles on dev and CI will test the other profiles.