Set up trust system, check it in peer discovery
This sets up the new trust system, and checks it in the peer discovery code, but
there isn't any code that distrusts peers yet. So this is technically a
regression: you can't be banned at the peer discovery layer anymore. It leads
into the behavior we want though. Small PRs :)