Set up trust system, check it in peer discovery (#1650)
* Set up trust system, check it in peer discovery

This sets up the new trust system, and checks it in the peer discovery code, but
there isn't any code that distrusts peers yet. So this is technically a
regression: you can't be banned at the peer discovery layer anymore. It leads
into the behavior we want though. Small PRs :)

* make reformat

* Use Cory's Docker images to pick up new Kad

* Actually use the right images

* Don't use non-existent tracetool image

* Hopefully use Docker images with the new kad executable

* Use Cory's Rust image

* Use Docker images with new async_kernel