Broadcast_pipe for transition frontiers (#1691)
* Shared_mvar for transition frontiers

Shared_mvar wraps an 'a option mvar -- in this case a
`Transition_frontier.t option` -- this way we can get events on None and Some.

The mvar is only consumed by Shared_mvar, no other code can assume the
inhabitedness of the mvar.

* Change to broadcast_pipe interface

All pipe listening functions also take a `close:unit Deferred.t` signal
to inform the underlying broadcast_pipe that it should close just the
pipe that started the iter/fold/etc.

* Remove printfs

* Rename to broadcast pipe more

* Requested changes

* Broadcast_pipe is synchronous. Better mirrors pipe

Broadcast_pipe is a pipe. It created strict pipes under the hood. It's
always synchronous.

* Remove backtrace elide false