No recent blocks check should obey stake semantics (#1747)
In integration tests, we need to make sure to handle the case where
all nodes lose some `i < c` slots in a row. This can and does happen
frequently and comes up specifically when the slot time is pushed a bit
low.

In the "no recent blocks", check we should also wait an extra `c-1`
slots before reporting an error.

Here we're using `c = 1` for sig as sig does expect a block to be
created every slot (and therefore doesn't change the old check).