Use best tip path changes in transaction pool (#1800)
I modified the transition frontier extension interface to expose changes in the
best tip path, modified the best tip diff view to expose changes in the user
commands, and used the data in the transaction pool. This should do the right
thing when the best tip change involves backtracking now.