Integration test tracing (#1845)
the new meaning of CODA_TRACING is a directory to put
trace files into. it will be created if it doesn't yet
exist. anything using Coda_worker will support it. the
env var is read by the thing that spawns the workers,
not the worker itself. trace_dir becomes a part of
worker configuration.