Nav closer to spec
It feels right to me to inline the top-level flex-logic in the jsx.
Honestly, the only reason I didn't inline everything is because I needed
to use that crazy checkbox hack and refer to another class. Let me know
what you think re: when we should be inlining styles.

Responsive reflowing shown here: