Refactor Syncable_ledger and add comments
I cleaned up some weird stuff and confusing names, and added a bunch of doc
comments on stuff that was confusing. I also changed it so the pipe of sync
query responses includes the associated queries. This prevents peers from
answering questions we didn't ask, and lets me remove data from the responses
that is contained in the associated queries.