Make coinbase an unsigned amount in Transaction snark (#1902)
In a coinbase transaction, coinbase amount is bound to the fee (signed) and coinbase fee (from the fee transfer) to the amount field of Transaction_union.t. This makes the effective coinbase-reward a signed value in checked computations. This means, when a coinbase amount from the transaction snark is required elsewhere, we either have to make coinbase amount a signed value through out or extract just the magnitude of the signed value

Changes:

    For a coinbase transaction, swapped the amount and fee values.
    Also added a new case "Invalid_proof" in staged ledger error